% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDataSet-as_list.R
\name{as_list}
\alias{as_list}
\title{Method to convert \code{MultiDataSet} to a list of matrices.}
\usage{
as_list(object)
}
\arguments{
\item{object}{\code{MultiDataSet} that will be filled.}
}
\value{
A lsit of matrices, one per dataset
}
\description{
Method to convert \code{MultiDataSet} to a list of matrices.
}

