% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGEO.R
\name{downloadGEO}
\alias{downloadGEO}
\title{Creation of a MultiDataSet with content of GEO GSE tags}
\usage{
downloadGEO(gse_list, gse_type)
}
\arguments{
\item{gse_list}{String vector with the GSE tags.}

\item{gse_type}{String vector with the GES data types (allowed:
code{"expression"}, \code{"snp"} and \code{"methylation"}).}
}
\description{
Given a vector of GSE tags with a single datasets. The function downloads
the dataset. It converts the downloaded \code{ExpressionSet} to the
correct set (\code{ExpressionSet} for expression data, \code{SnpSet} for
SNP data and \code{MethylationSet} for methylation data) and it
creates a MultiDataSet with them.
}
\examples{
\dontrun{
downloadGEO(
	gse_list = c("GSE40576", "GSE40732"), 
	gse_type = c("methylation", "expression")
)
}
}

