% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{getAssociation}
\alias{getAssociation}
\title{Method to extrat feature result from a ResultSet}
\usage{
getAssociation(object, rid, coef = 2, contrasts = NULL, sort = TRUE,
  fNames = c("chromosome", "start", "end", "genesymbol"), ...)
}
\arguments{
\item{object}{A \code{\link{ResultSet}} object.}

\item{rid}{The name or index of the result to be extracted.}

\item{coef}{(default \code{2}) Index of the coefficient to be extracted.}

\item{sort}{(default \code{TRUE}) If \code{TRUE}, results are ordered
by P-Value.}

\item{contrast}{(default \code{1}) When \code{code} corresponds to a
multicategorical variable, contasr selects the comparison.}
}
\value{
A \code{data.frame} with the result of the association study,
including P-Value and Fold Change.
}
\description{
Homologous methods from \code{limma}, \code{getAssociation} resturns a
\code{data.frame} with the \code{logFC} and  \code{PValue} per
featrue for the selcted \code{coef} and for given result (\code{rid}).
}
\examples{
data(gexp_r)
data(exp_r)
rst <- assocES(exp_r, gexp_r, formula=~sex+age)
getAssociation(rst, rid=1)
}
