% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambdaClayton.R
\name{lambdaClayton}
\alias{lambdaClayton}
\title{Lambda Calculation for a vector of P-Values}
\usage{
lambdaClayton(x, trim = 0.5)
}
\arguments{
\item{x}{Vector of P-Value}

\item{trim}{(default \code{0.5})}
}
\value{
A lambda value, indicating the inflation/deflation of the
analysis.
}
\description{
Implementation of Clayton's lambda score for a vector of P-Values
}
\examples{
lambdaClayton(runif(30))
}
\author{
Juan R. Gonzalez
}
