% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversionFunctions.R
\name{mae2mds}
\alias{mae2mds}
\title{Convert a \code{MultiAssayExperiment} to a \code{MultiDataSet}}
\usage{
mae2mds(MAE, warnings = TRUE)
}
\arguments{
\item{MAE}{a \code{MultiAssayExperiment}}

\item{warnings}{Logical to indicate if warnings will be displayed.}
}
\value{
\code{MultiDataSet} with the of the incoming \code{MultiAssayExperiment}.
}
\description{
This function creates a \code{MultiDataSet} using the data of a \code{MultiAssayExperiment}.
}

