% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversionFunctions.R
\name{mds2mae}
\alias{mds2mae}
\title{Convert a \code{MultiDataSet} to a \code{MultiAssayExperiment}}
\usage{
mds2mae(MDS)
}
\arguments{
\item{MDS}{a \code{MultiDataSet}}
}
\value{
\code{MultiAssayExperiment} with the of the incoming \code{MultiDataSet}.
}
\description{
This function creates a \code{MultiAssayExperiment} using the data of a \code{MultiDataSet}.
}

