% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDataSet.R
\docType{package}
\name{MultiDataSet}
\alias{MultiDataSet}
\alias{MultiDataSet-package}
\title{MultiDataSet: Implementation of the BRGE's basic classes}
\description{
Implementation of the BRGE's (Bioinformatic Research Group in Epidemiology from
 Center for Research in Environmental Epidemiology) MultiDataSet and MethylationSet. MultiDataSet
 is designed for integrating multi omics data sets and MethylationSet to contain normalized methylation data. MultiDataSet for integrating multi omics data sets
}
\seealso{
\linkS4class{MultiDataSet}
}

