% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ResultSet-ClassGenerics.R
\name{opt}
\alias{opt}
\title{Method to get the options sued to create the ResultSet}
\usage{
opt(object)
}
\arguments{
\item{object}{A \code{\link{ResultSet}} object.}
}
\value{
A list with the options used to create the \code{ResultSet}.
}
\description{
Method that returns a list with the options used to create the
\code{ResultSet}.
}
\examples{
data(rset)
opt(rset)
}
