% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{Function to draw a QQ Plot from a vector of numbers}
\usage{
qq_plot(values, show.lambda = TRUE)
}
\arguments{
\item{values}{Numeric vector of P.Values}

\item{show.lambda}{(default: \code{TRUE}) If \code{TRUE} shows lambda
score for the given model.}
}
\value{
An object obtained from \link[ggplot2]{ggplot}.
}
\description{
Function to draw a QQ Plot from a vector of numbers
}
\examples{
data(rset)
rst <- getAssociation(rset, rid = 1, fNames = NULL)
qq_plot(rst$P.Value)
}
