% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_AllClasses.R, R/ResultSet-fData.R,
%   R/ResultSet-length.R, R/ResultSet-names.R, R/ResultSet-opt.R,
%   R/ResultSet-plotAssociation.R, R/ResultSet-rid.R, R/ResultSet-topTable.R,
%   R/create_resultset.R
\docType{class}
\name{ResultSet}
\alias{ResultSet}
\alias{ResultSet-class}
\alias{fData,ResultSet-method}
\alias{length,ResultSet-method}
\alias{names,ResultSet-method}
\alias{opt,ResultSet-method}
\alias{plotAssociation,ResultSet-method}
\alias{rid,ResultSet-method}
\alias{topTable,ResultSet-method}
\alias{create_resultset}
\alias{ResultSet-methods}
\title{Class ResultSet}
\usage{
\S4method{fData}{ResultSet}(object)

\S4method{length}{ResultSet}(x)

\S4method{names}{ResultSet}(x)

\S4method{opt}{ResultSet}(object)

\S4method{plotAssociation}{ResultSet}(object, rid = 1, coef = 2,
  contrast = 1, type = c("manhattan", "qq", "volcano"), tPV, tFC,
  show.effect = FALSE)

\S4method{rid}{ResultSet}(object)

\S4method{topTable}{ResultSet}(object, rid, coef = 2, sort = TRUE)

create_resultset(fOrigin, lResults, fData, lOptions = list())
}
\arguments{
\item{object}{A \code{ResultSet} object.}

\item{x}{A \code{ResultSet} object.}

\item{rid}{Name or index of the internal result to be used}

\item{coef}{Coefficient to be returne, usually 2}

\item{contrast}{If coefficient to be used was multicategorical, number
of the contrast to be returned.}

\item{type}{Type of plot to be drawn}

\item{tPV}{Threshold for P-Value}

\item{tFC}{Threshold for log FC of effect}

\item{sort}{Indicates of the result should be ordered by P-Value}
}
\value{
An object of class \code{ResultSet}
}
\description{
Class \code{ResultSet} encapsulates the results from the association and
integration processes developed in \code{omicRexposome}. This include
the functions \code{\link{association}} and \code{\link{crossomics}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{fData}: Returns \code{data.frame} with feature's data.

\item \code{length}: Returns the amoung of analyses stored in the
\code{ResultSet}.

\item \code{names}: Returns the names of the omics data used to create
the \code{ResultSet}.

\item \code{opt}: Returns a list with the options used to create the
\code{ResultSet}

\item \code{plotAssociation}: Allows to plot a series of plots (QQ plot, Manhattan
plot and Volcano plot) depending on the results stored in the
\code{ResultSet}.

\item \code{rid}: Returns the names (aka. exposures) of the analyses
stored in the \code{ResultSet}.

\item \code{topTable}: Getter to obtain the raw \code{data.frame} from
association and integration analysis.
}}

\section{Slots}{

\describe{
\item{\code{fun_origin}}{Character containing the function that creates the object.}

\item{\code{results}}{List containing the results of the association/integration.}

\item{\code{fData}}{List containing the feature-data of the original objects.}

\item{\code{options}}{list of options used to create the \code{ResultSet}.}
}}

\examples{
create_resultset("hello", list(), list(), list())
}
