% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rset.R
\docType{data}
\name{rset}
\alias{rset}
\title{Example \code{ResultSet}}
\format{\code{ResultSet}}
\usage{
rset
}
\description{
Example \code{ResultSet} used in the functions examples and in the tests. The 
script used to generate it can be found in inst/scripts.
}
\keyword{datasets}
