% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{w_iclusterplus}
\alias{w_iclusterplus}
\title{Apply iClusterPlus clustering method to a MultiDataSet object}
\usage{
w_iclusterplus(object, commonSamples = TRUE, ...)
}
\arguments{
\item{object}{\code{MultiDataSet}}

\item{commonSamples}{Logical to indicate if common samples are selected}

\item{...}{Arguments passed to function \link[iClusterPlus]{iClusterPlus}}
}
\value{
A list of results from \link[iClusterPlus]{iClusterPlus}
}
\description{
Method \link[iClusterPlus]{iClusterPlus} is applied on a \link{MultiDataSet} object after
getting the common samples along all the contained datasets.
}
\note{
Argument \code{type} for \link[iClusterPlus]{iClusterPlus} is filled within the method.
}
