% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_MultiDataSet-ClassGenerics.R
\name{w_mcia}
\alias{w_mcia}
\title{Apply mcia integration method to a MultiDataSet object}
\usage{
w_mcia(object, ...)
}
\arguments{
\item{object}{\code{MultiDataSet}}

\item{...}{Arguments passed to function \link[omicade4]{mcia}}
}
\value{
A list of results from \link[omicade4]{mcia}
}
\description{
Method \link[omicade4]{mcia} is applied on a \link{MultiDataSet} object after
getting the common samples along all the contained datasets.
}
