% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_AllGenerics.R
\name{opt}
\alias{opt}
\title{Method to get the options sued to create the ResultSet}
\usage{
opt(object)
}
\arguments{
\item{object}{A \code{\link{ResultSet}} object.}
}
\value{
A list with the options used to create the \code{ResultSet}.
}
\description{
Method that returns a list with the options used to create the
\code{ResultSet}.
}
\examples{
data(gexp_r)
data(exp_r)
rst <- assocES(exp_r, gexp_r, formula=~sex+age)
opt(rst)
}
