% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_ResultSet-ClassGenerics.R,
%   R/ResultSet-fData.R, R/ResultSet-getAssociation.R, R/ResultSet-length.R,
%   R/ResultSet-names.R, R/ResultSet-opt.R, R/ResultSet-plot.R,
%   R/ResultSet-varLabels.R, R/create_resultset.R
\docType{class}
\name{ResultSet}
\alias{ResultSet}
\alias{ResultSet-class}
\alias{fData,ResultSet-method}
\alias{getAssociation,ResultSet-method}
\alias{length,ResultSet-method}
\alias{names,ResultSet-method}
\alias{opt,ResultSet-method}
\alias{plot,ResultSet,ANY-method}
\alias{varLabels,ResultSet-method}
\alias{create_resultset}
\alias{ResultSet-methods}
\title{Class ResultSet}
\usage{
\S4method{fData}{ResultSet}(object)

\S4method{getAssociation}{ResultSet}(object, rid, coef = 2,
  contrasts = NULL, sort = TRUE, fNames = c("chromosome", "start", "end",
  "genesymbol"), ...)

\S4method{length}{ResultSet}(x)

\S4method{names}{ResultSet}(x)

\S4method{opt}{ResultSet}(object)

\S4method{plot}{ResultSet,ANY}(x, y, rid = 1, coef = 2, contrast = NULL,
  type, tFC = 2, tPV = -log10(0.001), show.effect = FALSE,
  fNames = c("chromosome", "start"), ...)

\S4method{varLabels}{ResultSet}(object)

create_resultset(fOrigin, lResults, fData, lOptions = list())
}
\arguments{
\item{object}{A \code{ResultSet} object.}

\item{rid}{Name or index of the internal result to be used}

\item{coef}{Coefficient to be returne, usually 2}

\item{sort}{Indicates of the result should be ordered by P-Value}

\item{x}{A \code{ResultSet} object.}

\item{contrast}{If coefficient to be used was multicategorical, number
of the contrast to be returned.}

\item{type}{Type of plot to be drawn}

\item{tFC}{Threshold for log FC of effect}

\item{tPV}{Threshold for P-Value}
}
\value{
An object of class \code{ResultSet}
}
\description{
Class \code{ResultSet} encapsulates the results from the association and
integration processes developed in \code{omicRexposome}. This include
the functions \code{\link{association}} and \code{\link{crossomics}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{fData}: Returns \code{data.frame} with feature's data.

\item \code{getAssociation}: Getter to obtain the raw \code{data.frame} from
association and integration analysis.

\item \code{length}: Returns the amoung of analyses stored in the
\code{ResultSet}.

\item \code{names}: Returns the names of the omics data used to create
the \code{ResultSet}.

\item \code{opt}: Returns a list with the options used to create the
\code{ResultSet}

\item \code{plot}: Allows to plot a series of plots (QQ plot, Manhattan
plot and Volcano plot) depending on the results stored in the
\code{ResultSet}.

\item \code{varLabels}: Returns the names of the variables of the models used in
a \code{ResultSet}.
}}

\section{Slots}{

\describe{
\item{\code{fun_origin}}{Character containing the function that creates the object.}

\item{\code{results}}{List containing the results of the association/integration.}

\item{\code{fData}}{List containing the feature-data of the original objects.}

\item{\code{options}}{list of options used to create the \code{ResultSet}.}
}}

\examples{
create_resultset("hello", list(), list(), list())
}
